# [ISSUE-450] Fix performance regression