When compare TPC-DS 3T between main branch and branch-0.1, we found that round-robin is slower than branch-0.1, which is unexpected because the high-level allocation algorithm are basically the same.
```
main roundrobin:   5332s
branch-0.1:        5027s
```
After digging deeper I found that's because branch-0.1 first allocates master locations round-robin, then slave locations round-robin, however in main branch it allocates (master, slave) pairs round-robin. As a result, say one worker has two disks disk1 and disk2, we find all master partitions are allocated on disk1 and all slave partitions are allocated on disk2, which is different from branch-0.1 which disk1 and disk2 have the same number of both master partitions and slave partitions.
Experiments show that when change main branch algorithm vivic to branch-0.1, we get the performance back.
Time of q74:
```
branch-0.1:          58.749s
main before fix:     70.114s
main after fix:      58.987s
```

# [BUG]/[FEATURE] title

### What changes were proposed in this pull request?


### Why are the changes needed?

<!--
Please clarify why the changes are needed. For instance,
  1. If you propose a new API, clarify the use case for a new API.
  2. If you fix a bug, you can clarify why it is a bug.
-->

### What are the items that need reviewer attention?


### Related issues.
close #450 

### Related pull requests.


### How was this patch tested?


/cc @related-reviewer

/assign @main-reviewer